Adjusted the page structure.
