+++ web/app/auth/signin/page.js
@@ -1,50 +1,46 @@
 import * as React from 'react';
 import { SignInPage } from '@toolpad/core/SignInPage';
 import { AuthError } from 'next-auth';
-import { providerMap, signIn } from '../auth';
+import { providerMap, signIn } from '../../auth';
 
 export default function SignIn() {
     return (
-        <html lang="zh-cn">
-            <body>
-                <SignInPage
-                    providers={providerMap}
-                    signIn={async (
-                        provider,
-                        formData,
-                        callbackUrl,
-                    ) => {
-                        'use server';
-                        try {
-                            return await signIn(provider.id, {
-                                redirectTo: callbackUrl ?? '/',
-                            });
-                        } catch (error) {
-                            // The desired flow for successful sign in in all cases
-                            // and unsuccessful sign in for OAuth providers will cause a `redirect`,
-                            // and `redirect` is a throwing function, so we need to re-throw
-                            // to allow the redirect to happen
-                            // Source: https://github.com/vercel/next.js/issues/49298#issuecomment-1542055642
-                            // Detect a `NEXT_REDIRECT` error and re-throw it
-                            if (error instanceof Error && error.message === 'NEXT_REDIRECT') {
-                                throw error;
-                            }
-                            // Handle Auth.js errors
-                            if (error instanceof AuthError) {
-                                return {
-                                    error: error.message,
-                                    type: error.type,
-                                };
-                            }
-                            // An error boundary must exist to handle unknown errors
-                            return {
-                                error: 'Something went wrong.',
-                                type: 'UnknownError',
-                            };
-                        }
-                    }}
-                />
-            </body>
-        </html>
+        <SignInPage
+            providers={providerMap}
+            signIn={async (
+                provider,
+                formData,
+                callbackUrl,
+            ) => {
+                'use server';
+                try {
+                    return await signIn(provider.id, {
+                        redirectTo: callbackUrl ?? '/',
+                    });
+                } catch (error) {
+                    // The desired flow for successful sign in in all cases
+                    // and unsuccessful sign in for OAuth providers will cause a `redirect`,
+                    // and `redirect` is a throwing function, so we need to re-throw
+                    // to allow the redirect to happen
+                    // Source: https://github.com/vercel/next.js/issues/49298#issuecomment-1542055642
+                    // Detect a `NEXT_REDIRECT` error and re-throw it
+                    if (error instanceof Error && error.message === 'NEXT_REDIRECT') {
+                        throw error;
+                    }
+                    // Handle Auth.js errors
+                    if (error instanceof AuthError) {
+                        return {
+                            error: error.message,
+                            type: error.type,
+                        };
+                    }
+                    // An error boundary must exist to handle unknown errors
+                    return {
+                        error: 'Something went wrong.',
+                        type: 'UnknownError',
+                    };
+                }
+            }}
+        />
     );
 }
